agents/bidi-demo-WIP/pyproject.toml
[project]
name = "bidi-demo"
version = "0.1.0"
description = "ADK Bidi-streaming demo application"
authors = [
    { name = "Google LLC" }
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "google-adk>=1.20.0",
    "google-cloud-aiplatform[adk]>=1.93.0",
    "fastapi>=0.115.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
extend = "../../../pyproject.toml"

[tool.ruff.lint]
# Ignores function complication rule
ignore = ["C901", "PLR0915"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]


[tool.ruff.lint.isort]
known-first-party = ["app", "bidi_demo"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.wheel.shared-data]
"app/static" = "share/bidi-demo/static"
"app/.env" = "share/bidi-demo/.env.template"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"